MainPage-feat-:create popup renderetion
