+++ clients/auth-service/src/components/App/App.component.js
@@ -3,11 +3,11 @@ import landingPage from "../Landing/Langing.component";
 import './app.style.css';
 
 function App({ injectTo }) {
-  const appMain = `<div id="main-page" class="main-page"/>`;
-  injectHTMLToElement(appMain, injectTo);
+  // const appMain = `<div id="main-page" class="main-page"/>`;
+  // injectHTMLToElement(appMain, injectTo);
   
-  const mainPage = document.getElementById('main-page');
-  landingPage({ injectTo: mainPage });
+  // const mainPage = document.getElementById('main-page');
+  landingPage({ injectTo: injectTo });
 }
 
 export default App;
