+++ clients/auth-service/src/components/CardHeader/CardHeader.style.css
@@ -1,11 +1,11 @@
 .card__header {
   background-color: var(--btn1-color);
   color: var(--bg-color);
-  padding: 1rem 5rem;
+  padding: 3rem 0rem;
   text-align: center;
-  width: 60%;
-  margin-left: 20%;
+  width: 20vw;
   border-radius: 0 0 2rem 2rem;
+  font-size: 3rem;
 }
 
 .header h2 {
