+++ clients/auth-service/src/components/Form/Login/LoginForm.component.js
@@ -1,4 +1,34 @@
-const form = document.querySelector("form"),
+import { injectHTMLToElement } from '../../../utils/dom.service';
+import './LoginForm.style.css';
+
+
+function LoginForm({ injectTo }) {
+  const FormTemplate = `
+  <form action="#">
+    <div class="field username-field">
+        <div class="input">
+        <input type="text" placeholder="Username">
+        </div>
+        <div class="error error-text">Username field cant be empty</div>
+        <div class="error2 invalid-text">Invalid letters inserted</div>
+    </div>
+    <div class="password-field">
+        <div class="input">
+        <input type="password" placeholder="Password">
+        </div>
+        <div class="error error-text">Password field cant be empty</div>
+    </div>
+    <input type="submit" value="Login">
+    <div class="btn-section">
+    <button class="resetpass" id="reset" >Forgot Password?</button>
+    <button class="sign-up" id="sign-up">Sign Up</button>
+  </div>
+</form>
+  `;
+
+  injectHTMLToElement(FormTemplate, injectTo);
+
+  const form = document.querySelector("form"),
 userNameField = form.querySelector(".username-field"),
 userNameInput = userNameField.querySelector("input"),
 passwordField = form.querySelector(".password-field"),
@@ -42,3 +72,6 @@ function landingPage(){
 function usernameLetterValid(username){
   return /^[A-Za-z0-9]*$/.test(username);
 }
+}
+
+export default LoginForm;
