--- clients/auth-service/src/components/Form/Sign-Up/SignUpForm.style.css
@@ -1,84 +1,0 @@
-.form {
-	padding: 2rem 4rem;	
-}
-
-.input-field {
-	margin-bottom: 1rem;
-	padding-bottom: 0.5rem;
-	position: relative;
-}
-
-.input-field label {
-	display: inline-block;
-	margin-bottom: 1rem;
-}
-
-.input-field input {
-	border: 2px solid var(--2);
-	border-radius: 0.8rem;
-	display: block;
-	font-family: inherit;
-	font-size: 0.8rem;
-	padding: 0.8rem;
-	width: 100%;
-	margin-bottom: 1rem;
-}
-
-.input-field input:focus {
-	outline: 0;
-	border-color: #777;
-}
-
-.input-field.successful input {
-	border-color: #2ecc71;
-}
-
-.input-field.error input {
-	border-color: #e74c3c;
-}
-
-.input-field i {
-	visibility: hidden;
-	position: absolute;
-	top: 40px;
-	right: 10px;
-}
-
-.input-field.successful i.checkcircle {
-	color: #2ecc71;
-	visibility: visible;
-}
-
-.input-field.error i.exclamationcircle {
-	color: #e74c3c;
-	visibility: visible;
-}
-
-.input-field small {
-	color: #e74c3c;
-	position: absolute;
-	bottom: 0;
-	left: 0;
-	visibility: hidden;
-}
-
-.input-field.error small {
-	visibility: visible;
-}
-
-.form button {
-	background-color: var(--btn-color);
-	border: none ;
-	border-radius:1rem;
-	color: var(--bg-color);
-	display: block;
-	margin-top: 1.5rem;
-	width: 100%;
-	cursor: pointer;
-	padding: 0.5rem;
-	text-align: center;
-}
-
-.form button:hover{
-	background-color: var(--btn2-hover);
-}
