+++ clients/auth-service/src/components/Form/SignUp/SignUpForm.style.css
@@ -1,0 +1,106 @@
+.form {
+	padding: 3rem 3rem;
+    width: 23vw;
+    display: flex;
+    flex-direction: column;
+}
+
+.input-field {
+	margin-bottom: 1rem;
+	padding-bottom: 0.5rem;
+	position: relative;
+}
+
+.input-field label {
+    display: table;
+    margin-bottom: 1rem;
+    font-size: 2rem;
+    color: #060505;
+}
+
+.input-field input {
+	border: 0.2rem solid var(--2);
+    border-radius: 2.8rem;
+    display: block;
+    font-family: inherit;
+    font-size: 1.5rem;
+    padding: 1.5rem;
+    width: 100%;
+    margin-bottom: 1rem;
+}
+
+.input-field input:focus {
+	outline: 0;
+	border-color: #777;
+}
+
+.input-field.successful input {
+	border-color: #2ecc71;
+}
+
+.input-field.error input {
+	border-color: #e74c3c;
+}
+
+.input-field i {
+	visibility: hidden;
+	position: absolute;
+	top: 40px;
+	right: 10px;
+}
+
+.input-field.successful i.fa-check-circle {
+	color: #2ecc71;
+	visibility: visible;
+}
+
+.input-field.error i.fa-exclamation-circle {
+	color: #e74c3c;
+	visibility: visible;
+}
+
+.input-field small {
+	color: #e74c3c;
+    position: absolute;
+    font-size: 1rem;
+    bottom: 0;
+    left: 0;
+    visibility: hidden;
+}
+
+.input-field.error small {
+	visibility: visible;
+}
+
+.form button {
+	background-color: var(--btn-color);
+    border-radius: 4rem;
+    cursor: pointer;
+    font-size: 3rem;
+    padding: 1rem 2.5rem;
+	width: 50%;
+    margin-left: 5vw;
+}
+
+.form button:hover{
+	background-color: var(--btn2-hover);
+}
+
+.btn-section{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    margin: 5vh;
+  }
+  .resetpass{
+    border-radius: 3rem 0rem 0rem 3rem;
+    color: #060505;
+    background: #f9f9f9;
+    box-shadow: inset 0 0 15px -5px black;
+  }
+  .sign-up{
+    border-radius: 0rem 3rem 3rem 0rem;
+    color: #060505;
+    background: #f9f9f9;
+    box-shadow: inset 0 0 15px -5px black;
+  }
