+++ clients/auth-service/src/components/Landing/Landing.style.css
@@ -7,10 +7,17 @@
     --2:#665B52;
     --3:#488BAA;
   }
-body {
-  margin:0;
-  background-color: var(--bg-color);
-  font-family: 'Montserrat', sans-serif; 
+.popup-header{
+  width: 100%;
+  height: 100%;
+}
+.popup-main{
+  width: 100%;
+  height: 100%;
+}
+.popup-footer{
+  width: 100%;
+  height: 100%;
 }
 h1 {
   letter-spacing: 0.8rem;
@@ -50,25 +57,22 @@ h1 {
   text-align:center;
   background-color: var(--bg-color);
 }
-
 .popup{
-  position:fixed;
+  position: fixed;
   background-color: var(--bg-color);
-  border: 1px solid black;
-  display:flex;
-  flex-direction:column;
-  width:40vh;
-  height:60vh;
-  border-radius:20px;
-  justify-content:center;
-  text-align:center;
-  font-family: 'Montserrat', sans-serif; 
-  font-weight:bold;
-  top:20%;
-  left:42%; 
-  visibility:hidden;
-  z-index:1;
-
+  display: flex;
+  flex-direction: column;
+  box-shadow: #060505 0rem 0rem 1rem 0rem;
+  width: 60vh;
+  height: 80vh;
+  top: 10vh;
+  padding: 0 5vh;
+  border-radius: 3rem;
+  text-align: center;
+  visibility: hidden;
+  z-index: 1;
+  justify-content: space-between;
+  align-items: center;
 }
 .show {
   visibility: visible;
@@ -80,8 +84,18 @@ h1 {
   -webkit-animation: fadeIn 1s;
   animation: fadeIn 1s;
 }
-
-
+.exit{
+  color: #665b52;
+  background: none;
+  border-radius: 1rem;
+  position: absolute;
+  top: 0vh;
+  right: 1vw;
+  box-shadow: none;
+}
+.exit:hover{
+  color:var(--bg-color)
+}
 /* Animation */
 .parallax > use {
   animation: move-forever 25s cubic-bezier(.55,.5,.45,.5)     infinite;
