--- clients/auth-service/src/components/Landing/popup.js
@@ -1,9 +1,0 @@
-
-function myFunction() {
-  var popup = document.getElementById("myPopup");
-  popup.classList.toggle("show");
-} 
- function closeFunc() {
-      window.location.reload();
-  }
-
