+++ clients/auth-service/src/index.js
@@ -1,6 +1,6 @@
 import App from './components/App/app.component';
 import './index.css';
 
-const root = document.getElementById('root');
+// const root = document.getElementById('root');
 
-window.addEventListener('load', () => App({ injectTo: root }));
+window.addEventListener('load', () => App({ injectTo: document.body }));
